# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This API Handle MELI - Sprint 1 - Grupo 2",
    "title": "MELI Bootcamp API - Sprint 1 - Grupo 2",
    "termsOfService": "https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones",
    "contact": {
      "name": "API Support",
      "url": "https://developers.mercadolibre.com.ar/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/api/v1/sellers": {
      "get": {
        "description": "getAll sellers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sellers"
        ],
        "summary": "List sellers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "post": {
        "description": "Create sellers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sellers"
        ],
        "summary": "Create Sellers",
        "parameters": [
          {
            "description": "SellerHandler to Create",
            "name": "SellerHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.CreateSellerRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/sellers/{id}": {
      "get": {
        "description": "Get the details of a Sellers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sellers"
        ],
        "summary": "Get Sellers",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Sellers to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "delete": {
        "description": "Delete Sellers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sellers"
        ],
        "summary": "Delete Sellers",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a Sellers to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a Sellers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sellers"
        ],
        "summary": "Update Sellers",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Sellers to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated Sellers details",
            "name": "Sellers",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.UpdateSellerRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/warehouses": {
      "get": {
        "description": "get warehouses",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Warehouses"
        ],
        "summary": "List warehouses",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/web_utils.response"
              }
            }
          }
        }
      },
      "post": {
        "description": "Create warehouses",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Warehouses"
        ],
        "summary": "Create warehouses",
        "parameters": [
          {
            "description": "warehouses to create",
            "name": "WarehouseHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.WarehouseRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/warehouses/{id}": {
      "get": {
        "description": "get one warehouses by id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Warehouses"
        ],
        "summary": "Get warehouses",
        "parameters": [
          {
            "type": "integer",
            "description": "WarehouseHandler ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "delete": {
        "description": "delete warehouses by id",
        "tags": [
          "Warehouses"
        ],
        "summary": "Delete warehouses",
        "parameters": [
          {
            "type": "integer",
            "description": "WarehouseHandler ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "description": "update warehouses",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Warehouses"
        ],
        "summary": "Update warehouses",
        "parameters": [
          {
            "type": "integer",
            "description": "WarehouseHandler ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "WarehouseHandler to update",
            "name": "WarehouseHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.WarehouseRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "dtos.CreateSellerRequestDTO": {
      "type": "object",
      "required": [
        "address",
        "cid",
        "company_name",
        "locality_id",
        "telephone"
      ],
      "properties": {
        "address": {
          "type": "string"
        },
        "cid": {
          "type": "integer"
        },
        "company_name": {
          "type": "string"
        },
        "locality_id": {
          "type": "string"
        },
        "telephone": {
          "type": "string"
        }
      }
    },
...
    "dtos.UpdateSellerRequestDTO": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "cid": {
          "type": "integer"
        },
        "company_name": {
          "type": "string"
        },
        "locality_id": {
          "type": "string"
        },
        "telephone": {
          "type": "string"
        }
      }
    },
    "dtos.WarehouseRequestDTO": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "minimum_capacity": {
          "type": "integer"
        },
        "minimum_temperature": {
          "type": "integer"
        },
        "telephone": {
          "type": "string"
        },
        "warehouse_code": {
          "type": "string"
        }
      }
    },
...
    "web_utils.response": {
      "type": "object",
      "properties": {
        "data": {}
      }
    }
  }
}