# internal/application/dtos/seller_create_request.go
package dtos

type CreateSellerRequestDTO struct {
	CID         int    `json:"cid" binding:"required"`
	CompanyName string `json:"company_name" binding:"required"`
	Address     string `json:"address" binding:"required"`
	Telephone   string `json:"telephone" binding:"required"`
	LocalityID  string `json:"locality_id" binding:"required"`
}